Resolve the parameters after loading the input model [main]
So that the tool can continue working even for parametric
models.

This uses the functionality of the robot-model-tools package,
as before.

TODO: it would be nice if the kingen toolchain (ilk-generator
and ilk-compiler) was able to deal with parametrized models
all the way to the generated code (instead of having to settle
the values at the beginning).