Introduce the keyword "_identity_" in ILK models
This change accounts for the corner cases when the
pose of a frame relative to itself is required by
the query.*

This fixes the previous behavior of using the
default pose identifier ('<frame>__<frame>'),
which eventually leads to an undefined reference.

Now the keyword "_identity_" is used for such poses,
in the generated ILK model.

Of course consumers of the ILK models (like the
ILK-compiler) must be adapted accordingly.

*: this case arises e.g. when requesting a Jacobian
whose reference is a joint frame