Deal with fixed joints in the input model
Requires a version of the robot-model-tools with the
explicit enum constant for fixed joints.

Generate an entry in the model constants for the fixed-joints
transforms, using the keyword '_identity_' as the value
  -> the ILK compiler must be aware of this new possibility

When generating the ILK model, consider any but fixed-joints
to be "valid", meaning fixed-joints will not be listed in the
"joints" section, nor in the "joint-poses" section.