Fix generation of wrong constant transforms
I was doing the usual dumb mistake of inverting an homogeneous
coordinate transform by simply taking the negative of the
translation vector.

Fixed by properly leveraging the API of kgprim.ct, which allows
to be explicit about the polarity of the desired transform:
for every joint-frame pose I query for both transforms, instead
of querying for only one and doing the inversion by hand.